requirements.txt
hikari==2.0.0.dev103
hikari-tanjun==2.1.2a1
python-dotenv==0.19.1
requests==2.26.0
